+++ medic-src/src/step/mod.rs
@@ -5,7 +5,7 @@ pub use shell_config::ShellConfig;
 pub use step_config::StepConfig;
 
 use crate::runnable::Runnable;
-use crate::AppResult;
+use crate::{AppResult, Check};
 use serde::Deserialize;
 use std::fmt;
 use std::process::{Command, Stdio};
@@ -13,6 +13,7 @@ use std::process::{Command, Stdio};
 #[derive(Debug, Deserialize)]
 #[serde(untagged)]
 pub enum Step {
+    Check(Check),
     Shell(ShellConfig),
     Step(StepConfig),
     Doctor(DoctorConfig),
@@ -29,33 +30,37 @@ pub struct DoctorConfig {
 impl Runnable for Step {
     fn allow_failure(&self) -> bool {
         match self {
+            Step::Check(config) => config.allow_failure(),
+            Step::Doctor(_) => false,
             Step::Shell(config) => config.allow_failure(),
             Step::Step(config) => config.allow_failure(),
-            Step::Doctor(_) => false,
         }
     }
 
     fn run(self) -> AppResult<()> {
         match self {
+            Step::Check(config) => config.run(),
+            Step::Doctor(_) => run_doctor(),
             Step::Shell(config) => config.run(),
             Step::Step(config) => config.run(),
-            Step::Doctor(_) => run_doctor(),
         }
     }
 
     fn to_command(&self) -> Option<Command> {
         match self {
+            Step::Check(config) => config.to_command(),
+            Step::Doctor(_) => doctor_command(),
             Step::Shell(config) => config.to_command(),
             Step::Step(config) => config.to_command(),
-            Step::Doctor(_) => doctor_command(),
         }
     }
 
     fn verbose(&self) -> bool {
         match self {
+            Step::Check(config) => config.verbose,
+            Step::Doctor(_) => true,
             Step::Shell(config) => config.verbose,
             Step::Step(config) => config.verbose,
-            Step::Doctor(_) => true,
         }
     }
 }
@@ -63,9 +68,10 @@ impl Runnable for Step {
 impl fmt::Display for Step {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         match self {
+            Step::Check(config) => config.fmt(f),
+            Step::Doctor(_) => write!(f, "\x1b[36m== Doctor ===\x1b[0m"),
             Step::Shell(config) => config.fmt(f),
             Step::Step(config) => config.fmt(f),
-            Step::Doctor(_) => write!(f, "\x1b[36m== Doctor ===\x1b[0m"),
         }
     }
 }